Fix invalid matrix elements being produced by parallel parser
Issue was due to the pos increment needing to be equal to the number of unique kmers, which you
can't know apriori from processing. The fix is to take the max number of unique kmers as the
chunk belonging to a sequence thread, then cleaning up the results by removing any invalid
kmers at the end (represented by index==-1).